move some code to a submod, fix `BitSet::from_bools()`
